add look up support

+++ index.js
@@ -14,7 +14,7 @@ export default function makeDefaultCacheFunctions ({ makeCache }) {
    * @param {number} expireAt - Timestamp in milliseoconds for removal
    * @returns {object}
    */
-  async function set ({ id, content, timeLeftS, expireAt, unique = true }) {
+  async function set ({ id, content, timeLeftS, expireAt, unique = true, lookUps = [] } = {}) {
     if (expireAt && timeLeftS) {
       throw new Error('Specify either timeLeftS or expireAt but not both')
     }
@@ -29,12 +29,20 @@ export default function makeDefaultCacheFunctions ({ makeCache }) {
     if (unique && await find({ id })) {
       throw new Error('Cache duplication error')
     }
-    const result = await client.set(
-      id,
-      content,
-      timeLeftS ? { EX: timeLeftS } : undefined
-    )
-    return result === 'OK' ? { id, content } : null
+    const results = await Promise.all([
+      client.set(
+        id,
+        content,
+        timeLeftS ? { EX: timeLeftS } : undefined
+      ),
+      ...lookUps.map(lookUp => client.set(
+        lookUp,
+        id,
+        timeLeftS ? { EX: timeLeftS } : undefined
+      ))
+    ])
+    const anyNotOk = results.some(result => result !== 'OK')
+    return anyNotOk ? null : { id, content }
   }
   /**
    * 
@@ -43,26 +51,46 @@ export default function makeDefaultCacheFunctions ({ makeCache }) {
    * @param {number} expireAt - Timestamp in milliseoconds for removal
    * @returns {object} Cached object
    */
-  async function setObj ({ info, timeLeftS, expireAt, unique = true }) {
+  async function setObj ({ info, timeLeftS, expireAt, unique = true, lookUps } = {}) {
     const result = await set({
       id: info.id,
       content: JSON.stringify(info),
       timeLeftS,
       expireAt,
-      unique
+      unique,
+      lookUps
     })
     return result ? info : null
   }
-  async function find (id) {
+  async function find ({ id, lookUp } = {}) {
+    if (id == null && lookUp == null) {
+      throw new Error('No id or lookUp supplied')
+    }
     const client = await makeCache()
+    if (id == null) {
+      id = await client.get(lookUp)
+      if (id == null) return null
+    }
     return client.get(id)
   }
-  async function findObj (id) {
-    const result = await find({ id })
+  async function findObj ({ id, lookUp } = {}) {
+    if (id == null && lookUp == null) {
+      throw new Error('No id or lookUp supplied')
+    }
+    const result = await find({ id, lookUp })
     return result ? JSON.parse(result) : null
   }
-  async function remove (id) {
+  async function remove ({ id, lookUp, lookUps = [] } = {}) {
+    if (id == null && lookUp == null) {
+      throw new Error('No id or lookUp supplied')
+    }
+    if (lookUp != null && (lookUps == null || lookUps.length === 0)) {
+      throw new Error('lookUp supplied but no lookUps defined')
+    }
     const client = await makeCache()
-    return client.del(id)
+    await Promise.all([
+      client.del(id),
+      ...lookUps.map(lookUp => client.del(lookUp))
+    ])
   }
 }
